+++ tests/testthat/test-rbackground.R
@@ -12,29 +12,35 @@ test_that("Test collecting LazyFrame in background", {
 
 test_that("Test using $map() in background", {
   # change capacity
-  pl$set_global_rpool_cap(0)
-  expect_equal(pl$get_global_rpool_cap(), list(available = 0, capacity = 0))
-  pl$set_global_rpool_cap(1)
-  expect_equal(pl$get_global_rpool_cap(), list(available = 0, capacity = 1))
-
+  pl$set_options(rpool_cap = 0)
+  expect_equal(pl$options$rpool_cap, 0)
+  expect_equal(pl$options$rpool_active, 0)
+  pl$set_options(rpool_cap = 1)
+  expect_equal(pl$options$rpool_cap, 1)
+  expect_equal(pl$options$rpool_active, 0)
 
 
   compute = lf$select(pl$col("y")$map(\(x) x * x, in_background = FALSE))
-  compute_bg = lf$select(pl$col("y")$map(\(x) x * x, in_background = TRUE))
+  compute_bg = lf$select(pl$col("y")$map(\(x) {
+    Sys.sleep(2)
+    x * x
+  }, in_background = TRUE))
   res_ref = compute$collect()$to_data_frame()
 
   # no process spawned yet
-  expect_equal(pl$get_global_rpool_cap(), list(available = 0, capacity = 1))
+  expect_equal(pl$options$rpool_cap, 1)
+  expect_equal(pl$options$rpool_active, 0)
 
   # process was spawned
   res_fg_map_bg = compute_bg$collect()$to_data_frame()
-  expect_equal(pl$get_global_rpool_cap(), list(available = 1, capacity = 1))
+  expect_equal(pl$options$rpool_cap, 1)
+  expect_equal(pl$options$rpool_active, 1)
 
   # same result
   expect_identical(res_ref, res_fg_map_bg)
 
   # cannot collect in background without a cap
-  pl$set_global_rpool_cap(0)
+  pl$set_options(rpool_cap = 0)
   handle = compute_bg$collect_in_background()
   res = result(handle$join())
   expect_rpolarserr(unwrap(res), c("When", "Hint", "PlainErrorMessage"))
@@ -42,7 +48,6 @@ test_that("Test using $map() in background", {
     res$err$contexts() |> tail(1) |> unlist(use.names = FALSE),
     "cannot run background R process with zero capacity"
   )
-
 
   # can print handle after exhausted
   handle |>
@@ -54,4 +59,29 @@ test_that("Test using $map() in background", {
 
   # gives correct err message
   expect_rpolarserr(handle$join(), "Handled")
+})
+
+
+test_that("reset rpool_cap", {
+  pl$reset_options()
+  orig = pl$options$rpool_cap
+  pl$set_options(rpool_cap = orig + 1)
+  expect_different(pl$options$rpool_cap, orig)
+  pl$reset_options()
+  expect_identical(pl$options$rpool_cap, orig)
+})
+
+
+test_that("rpool errors", {
+
+  ctx =  pl$set_options(rpool_cap = c(1, 2)) |> get_err_ctx()
+  expect_identical(ctx$BadArgument, "rpool_cap")
+  expect_true(startsWith(ctx$TypeMismatch,"i64"))
+
+  ctx = pl$set_options(rpool_cap = -1) |> get_err_ctx()
+  expect_identical(ctx$ValueOutOfScope, "cannot be less than zero")
+
+  ctx = {polars_optenv$rpool_active <- 0} |> get_err_ctx()
+  expect_true(endsWith(ctx$PlainErrorMessage,"rpool_active cannot be set directly"))
+
 })